--- app.js
@@ -1,6 +1,0 @@
-const server =require('server');
-const port=4004;
-
-server.listen(port,()=>{
-    console.log(`listening on port ${port}`);
-})
