--- client-server.md
@@ -1,31 +1,0 @@
-Client-server architecture is a model where computing resources and responsibilities are distributed between clients and servers in a network. Here's a breakdown of its key components:
-
-1. **Client**: 
-   - The client is typically a user-facing device or application that requests services or resources from a server.
-   - Examples include web browsers, mobile apps, desktop applications, or IoT devices.
-   - Clients initiate communication with servers by sending requests for data or services.
-
-2. **Server**:
-   - The server is a centralized system that provides resources, services, or data to clients.
-   - Servers handle client requests, process them, and send back the requested information or perform the requested actions.
-   - Examples include web servers, application servers, database servers, or cloud servers.
-
-3. **Communication**:
-   - Clients and servers communicate with each other over a network, typically using protocols such as HTTP, TCP/IP, or WebSocket.
-   - Clients initiate communication by sending requests to the server, and servers respond to these requests with the requested data or perform the requested actions.
-   - Communication can be synchronous, where the client waits for a response from the server before proceeding, or asynchronous, where the client continues its operation while waiting for a response.
-
-4. **Responsibilities**:
-   - In a client-server architecture, responsibilities are typically divided between clients and servers based on their capabilities and roles.
-   - Clients are responsible for presenting information to users, handling user input, and initiating requests for services or resources.
-   - Servers are responsible for processing client requests, executing business logic, accessing data, and providing responses to clients.
-
-5. **Scalability**:
-   - Client-server architecture allows for scalable systems where multiple clients can connect to a single server or multiple servers in a distributed environment.
-   - Servers can be scaled vertically by upgrading hardware resources or horizontally by adding more servers to handle increased load.
-
-6. **Security**:
-   - Security measures such as authentication, encryption, and access control are essential in client-server architecture to protect data and resources.
-   - Servers often implement security measures to authenticate clients, authorize access to resources, and secure communication channels.
-
-Overall, client-server architecture provides a flexible and scalable model for building distributed systems where clients and servers interact to fulfill various computing needs. It's widely used in web applications, mobile apps, cloud computing, and many other domains.
