--- results.csv
@@ -1,5 +1,0 @@
-Operation,Number 1,Number 2,Result
-Addition,2,6,8
-Subtraction,8,6,2
-Multiplication,4,7,28
-Division,22,11,2
