--- src/config/validationJoi.js
@@ -1,9 +1,0 @@
-const Joi = require("joi");
-const userSchema = Joi.object({
-  username: Joi.string().alphanum().min(3).max(30).required(),
-
-  email: Joi.string().email().required(),
-
-  password: Joi.string().pattern(new RegExp("^[a-zA-Z0-9]{3,30}$")).required(),
-});
-module.exports = userSchema;
