--- src/controllers/index.ts
@@ -1,5 +1,0 @@
-import promise from "./asyncController";
-import validateControl from "./validateControl";
-
-
-module.exports={promise,validateControl};
