--- src/controllers/validateControl.js
@@ -1,6 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const validateControl = (req, res) => {
-    res.send(req.query.name);
-};
-exports.default = validateControl;
