--- src/controllers/validateControl.ts
@@ -1,6 +1,0 @@
-import {Request,Response,NextFunction} from "express";
-const validateControl = (req:Request,res:Response) => {
-    res.send(req.query.name);
-  };
-  
-  export default validateControl;
