--- src/middleware/requestLogger.js
@@ -1,7 +1,0 @@
-export const requestLogger = (req, res, next) => {
-  console.log(`[${new Date().toISOString()}] ${req.method} ${req.url}`);
-  next(); 
-};
-
-
-
