--- src/routes/index.js
@@ -1,4 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const userRouter = require('./userRouter');
-exports.default = userRouter;
