--- src/routes/index.ts
@@ -1,5 +1,0 @@
- const userRouter = require('./userRouter');
-
-export default userRouter;
-
-
