--- validations.md
@@ -1,17 +1,0 @@
-Validations play a crucial role in security by helping to prevent various types of attacks and vulnerabilities. Here are several key reasons why validations are important for security:
-
-1. **Input Sanitization**: Validations help ensure that user input is properly sanitized before being processed by the application. This helps prevent injection attacks such as SQL injection, NoSQL injection, and Cross-Site Scripting (XSS) attacks, where malicious code is injected into input fields to manipulate or exploit the application.
-
-2. **Data Integrity**: Validations ensure that the data entered by users conforms to the expected format and structure. This helps maintain data integrity by preventing users from submitting invalid or malformed data, which could potentially corrupt the database or cause errors in the application.
-
-3. **Preventing Security Misconfigurations**: Validations help enforce security policies and prevent security misconfigurations by ensuring that input data adheres to predefined rules and restrictions. This includes enforcing password complexity requirements, limiting file upload sizes, and validating input against known attack patterns.
-
-4. **Preventing Authentication Bypass**: Validations are critical for ensuring the integrity of authentication mechanisms. By validating user input during authentication processes, such as login forms, applications can prevent common attacks like credential stuffing, brute force attacks, and username enumeration.
-
-5. **Preventing Business Logic Vulnerabilities**: Validations help enforce business logic rules and prevent security vulnerabilities that may arise from improper input handling. For example, validating input data before executing critical operations (such as financial transactions) can help prevent unauthorized access and data manipulation.
-
-6. **Mitigating Denial of Service (DoS) Attacks**: Validations can help mitigate the risk of Denial of Service (DoS) attacks by imposing limits on the size and frequency of incoming requests. By validating and rate-limiting user input, applications can prevent attackers from overwhelming the system with malicious or excessive requests.
-
-7. **Compliance Requirements**: Validations are often required to comply with security standards, regulations, and industry best practices. Many regulatory frameworks, such as the Payment Card Industry Data Security Standard (PCI DSS) and the General Data Protection Regulation (GDPR), mandate the implementation of adequate input validation controls to protect sensitive data and ensure user privacy.
-
-Overall, validations are a fundamental aspect of building secure software applications. By validating input data, enforcing security policies, and mitigating common attack vectors, developers can significantly reduce the risk of security breaches and protect both the application and its users from potential threats.
